+++ packages/combobox/src/Combobox/Combobox.styles.ts
@@ -1,3 +1,5 @@
+import { transparentize } from 'polished';
+
 import { css } from '@leafygreen-ui/emotion';
 import { createUniqueClassName, Theme } from '@leafygreen-ui/lib';
 import { palette } from '@leafygreen-ui/palette';
@@ -10,31 +12,69 @@ import {
   typeScales,
 } from '@leafygreen-ui/tokens';
 
-import { chipHeight } from '../Chip/Chip.styles';
+import { fontSize, getChipHeight, lineHeight } from '../Chip/Chip.styles';
 import { ComboboxSize as Size, Overflow } from '../Combobox.types';
 
 // Rename the variable defined in chip styles
-const inputHeight = chipHeight;
+const inputHeight = getChipHeight;
+
+// Gap between each chip
+const flexGap = 4;
 
 /**
- * Width of the widest character (in px)
+ * The min-height of the combobox.
  */
-export const maxCharWidth: Record<Size, number> = {
-  [Size.Default]: typeScales.body1.fontSize,
-  [Size.Large]: typeScales.body2.fontSize,
+export const wrapperHeight: Record<Size, number> = {
+  [Size.XSmall]: 22,
+  [Size.Small]: 28,
+  [Size.Default]: 36,
+  [Size.Large]: 48,
+};
+
+/**
+ * Util that calculates the Y padding.
+ * `(wrapperHeight - inputHeight(- (borderTop + borderBottom)) / 2`
+ */
+const getYPadding = (size: Size) => {
+  return (wrapperHeight[size] - inputHeight(size) - 2) / 2;
 };
 
 /**
  * Size of combobox x & y padding (in px)
+ * (wrapperHeight - inputHeight(- (borderTop + borderBottom)) / 2
  */
-export const comboboxPadding: Record<Size, { x: number; y: number }> = {
+export const comboboxPadding: Record<
+  Size,
+  {
+    y: number;
+    xLeftWithChip: number;
+    xLeftWithoutChip: number;
+    xRight: number;
+  }
+> = {
+  [Size.XSmall]: {
+    y: getYPadding(Size.XSmall),
+    xLeftWithChip: 1,
+    xLeftWithoutChip: 10,
+    xRight: 4,
+  },
+  [Size.Small]: {
+    y: getYPadding(Size.Small),
+    xLeftWithChip: 4,
+    xLeftWithoutChip: 10,
+    xRight: 8,
+  },
   [Size.Default]: {
-    y: (36 - inputHeight[Size.Default] - 2) / 2,
-    x: spacing[2] - 1,
+    y: getYPadding(Size.Default),
+    xLeftWithChip: 6,
+    xLeftWithoutChip: 12,
+    xRight: 12,
   },
   [Size.Large]: {
-    y: (48 - inputHeight[Size.Large] - 2) / 2,
-    x: spacing[2] - 1,
+    y: getYPadding(Size.Large),
+    xLeftWithChip: spacing[2] - 1,
+    xLeftWithoutChip: spacing[2] - 1,
+    xRight: spacing[2] - 1,
   },
 };
 
@@ -44,34 +84,23 @@ export const clearButtonIconSize = 28;
 /** Width of the dropdown caret icon (in px) */
 export const caretIconSize = spacing[3];
 
-const minWidth: Record<Size, number> = {
-  [Size.Default]:
-    maxCharWidth[Size.Default] +
-    2 * comboboxPadding[Size.Default].x +
-    caretIconSize +
-    2, // + 2 for border: ;
-  [Size.Large]:
-    maxCharWidth[Size.Large] +
-    2 * comboboxPadding[Size.Large].x +
-    caretIconSize +
-    2, // + 2 for border: ;
-};
-
 export const chipClassName = createUniqueClassName('combobox-chip');
 
 export const comboboxParentStyle = (size: Size): string => {
   return css`
     font-family: ${fontFamilies.default};
     width: 100%;
-    min-width: ${minWidth[size]}px;
+    min-width: ${fontSize[size] +
+    2 * comboboxPadding[size].xLeftWithChip +
+    caretIconSize +
+    2}px;
   `;
 };
 
 export const baseComboboxStyles = css`
-  display: grid;
-  grid-auto-flow: column;
-  grid-template-columns: 1fr ${caretIconSize}px;
+  display: flex;
   align-items: center;
+  gap: ${spacing[2]}px;
   cursor: text;
   transition: ${transitionDuration.default}ms ease-in-out;
   transition-property: background-color, box-shadow, border-color;
@@ -79,6 +108,23 @@ export const baseComboboxStyles = css`
   width: 100%;
   max-width: 100%;
   border-radius: 6px;
+  position: relative;
+  overflow: hidden;
+
+  // Overflow shadow
+  ::after {
+    content: '';
+    position: absolute;
+    width: 100%;
+    height: 20px;
+    bottom: -21px;
+    left: 50%;
+    translate: -50% 0%;
+    border-radius: 20%;
+    box-shadow: 0 0 0 0 rgb(255 255 255 / 0%);
+    transition: ${transitionDuration.default}ms linear;
+    transition-property: box-shadow;
+  }
 `;
 
 export const comboboxThemeStyles: Record<Theme, string> = {
@@ -102,8 +148,16 @@ export const comboboxThemeStyles: Record<Theme, string> = {
   `,
 };
 
-export const comboboxSizeStyles = (size: Size) => css`
-  padding: ${comboboxPadding[size].y}px ${comboboxPadding[size].x}px;
+export const comboboxSizeStyles = (
+  size: Size,
+  isMultiselectWithSelections: boolean,
+) => css`
+  padding-top: ${comboboxPadding[size].y}px;
+  padding-bottom: ${comboboxPadding[size].y}px;
+  padding-left: ${isMultiselectWithSelections
+    ? `${comboboxPadding[size].xLeftWithChip}px`
+    : `${comboboxPadding[size].xLeftWithoutChip}px`};
+  padding-right: ${comboboxPadding[size].xRight}px;
 `;
 
 export const comboboxDisabledStyles: Record<Theme, string> = {
@@ -134,22 +188,36 @@ export const comboboxFocusStyle: Record<Theme, string> = {
   [Theme.Light]: css`
     &:focus-within {
       border-color: transparent;
-      background-color: ${palette.white};
       box-shadow: ${focusRing[Theme.Light].input};
     }
   `,
   [Theme.Dark]: css`
     &:focus-within {
       border-color: transparent;
-      background-color: ${palette.gray.dark4};
       box-shadow: ${focusRing[Theme.Dark].input};
     }
   `,
 };
 
-export const comboboxSelectionStyles = css`
-  grid-template-columns: 1fr ${clearButtonIconSize}px ${caretIconSize}px;
+export const iconsWrapperBaseStyles = css`
+  display: flex;
+  align-items: center;
 `;
+
+export const iconsWrapperSizeStyles: Record<Size, string> = {
+  [Size.XSmall]: css`
+    gap: ${spacing[1]}px;
+  `,
+  [Size.Small]: css`
+    gap: ${spacing[2]}px;
+  `,
+  [Size.Default]: css`
+    gap: ${spacing[2]}px;
+  `,
+  [Size.Large]: css`
+    gap: ${spacing[2]}px;
+  `,
+};
 
 export const inputWrapperStyle = ({
   overflow,
@@ -168,11 +236,12 @@ export const inputWrapperStyle = ({
       return css`
         ${baseWrapperStyle}
         display: block;
-        height: ${inputHeight[size]}px;
+        height: ${inputHeight(size)}px;
         white-space: nowrap;
         overflow-x: scroll;
         scroll-behavior: smooth;
         scrollbar-width: none;
+        line-height: 1;
 
         &::-webkit-scrollbar {
           display: none;
@@ -198,9 +267,10 @@ export const inputWrapperStyle = ({
         ${baseWrapperStyle}
         display: flex;
         flex-wrap: wrap;
-        gap: 4px;
+        gap: ${flexGap}px;
         overflow-x: hidden;
-        min-height: ${inputHeight[size]}px;
+        min-height: ${inputHeight(size)}px;
+        max-height: calc((${getChipHeight(size) * 3}px) + (${flexGap}px * 2));
       `;
     }
   }
@@ -217,6 +287,12 @@ export const baseInputElementStyle = css`
   padding: 0;
   margin: 0;
   text-overflow: ellipsis;
+  vertical-align: top;
+
+  // Only add padding if there are chips
+  &:not(:first-child) {
+    padding-left: ${spacing[1]}px;
+  }
 
   &:placeholder-shown {
     min-width: 100%;
@@ -239,27 +315,25 @@ export const inputElementThemeStyle: Record<Theme, string> = {
   `,
 };
 
-export const inputElementSizeStyle: Record<Size, string> = {
-  [Size.Default]: css`
-    height: ${inputHeight[Size.Default]}px;
-    font-size: ${typeScales.body1.fontSize}px;
-    line-height: ${typeScales.body1.lineHeight}px;
-    min-width: ${maxCharWidth[Size.Default]}px;
-    // Only add padding if there are chips
-    &:not(:first-child) {
-      padding-left: 4px;
+export const inputElementDisabledThemeStyle: Record<Theme, string> = {
+  [Theme.Light]: css`
+    &::placeholder {
+      color: ${palette.gray.dark1};
     }
   `,
-  [Size.Large]: css`
-    height: ${inputHeight[Size.Large]}px;
-    font-size: ${typeScales.body2.fontSize}px;
-    line-height: ${typeScales.body2.lineHeight}px;
-    min-width: ${maxCharWidth[Size.Large]}px;
-    &:not(:first-child) {
-      padding-left: 6px;
+  [Theme.Dark]: css`
+    &::placeholder {
+      color: ${palette.gray.dark1};
     }
   `,
 };
+
+export const inputElementSizeStyle = (size: Size) => css`
+  height: ${inputHeight(size)}px;
+  font-size: ${fontSize[size]}px;
+  line-height: ${lineHeight[size]}px;
+  min-width: ${fontSize[size]}px;
+`;
 
 export const inputElementTransitionStyles = (isOpen: boolean) => css`
   /*
@@ -277,7 +351,7 @@ export const multiselectInputElementStyle = (
 ) => {
   const inputLength = inputValue?.length ?? 0;
   return css`
-    width: ${inputLength * maxCharWidth[size]}px;
+    width: ${inputLength * fontSize[size]}px;
     max-width: 100%;
   `;
 };
@@ -285,12 +359,12 @@ export const multiselectInputElementStyle = (
 export const clearButtonStyle = css`
   // Add a negative margin so the button takes up the same space as the regular icons
   margin-block: calc(${caretIconSize / 2}px - 100%);
+  margin-inline: -6px;
 `;
 
-export const endIconStyle = (size: Size) => css`
+export const endIconStyle = css`
   height: ${caretIconSize}px;
   width: ${caretIconSize}px;
-  margin-inline-end: calc(${comboboxPadding[size].x}px / 2);
 `;
 
 export const errorMessageThemeStyle: Record<Theme, string> = {
@@ -302,18 +376,48 @@ export const errorMessageThemeStyle: Record<Theme, string> = {
   `,
 };
 
-export const errorMessageSizeStyle: Record<Size, string> = {
-  [Size.Default]: css`
-    font-size: ${typeScales.body1.fontSize}px;
-    line-height: ${typeScales.body1.lineHeight}px;
-    padding-top: ${comboboxPadding[Size.Default].y}px;
+export const errorMessageSizeStyle = (size: Size) => css`
+  font-size: ${fontSize[size]}px;
+  line-height: ${lineHeight[size]}px;
+  padding-top: ${comboboxPadding[size].y}px;
+`;
+
+export const labelDescriptionContainerStyle = css`
+  margin-bottom: ${spacing[1]}px;
+  display: flex;
+  flex-direction: column;
+`;
+
+export const labelDescriptionLargeStyles = css`
+  font-size: ${typeScales.large.fontSize}px;
+  line-height: ${typeScales.large.lineHeight}px;
+`;
+
+export const comboboxOverflowShadowStyles: Record<Theme, string> = {
+  [Theme.Light]: css`
+    ::after {
+      box-shadow: 0px 0px 7px 5px ${transparentize(0.85, palette.black)};
+    }
   `,
-  [Size.Large]: css`
-    font-size: ${typeScales.body2.fontSize}px;
-    line-height: ${typeScales.body2.lineHeight}px;
-    padding-top: ${comboboxPadding[Size.Large].y}px;
+  [Theme.Dark]: css`
+    ::after {
+      width: 95%;
+      box-shadow: 0px -7px 12px 5px rgb(0 0 0 / 50%);
+    }
   `,
 };
-export const labelDescriptionContainerStyle = css`
-  margin-bottom: 2px;
-`;
+
+export const errorIconThemeStyles: Record<Theme, string> = {
+  [Theme.Light]: palette.red.base,
+  [Theme.Dark]: palette.red.light1,
+};
+
+export const caretIconThemeStyles: Record<Theme, string> = {
+  [Theme.Light]: palette.gray.dark2,
+  [Theme.Dark]: palette.gray.light1,
+};
+
+export const caretIconDisabledStyles: Record<Theme, string> = {
+  [Theme.Light]: palette.gray.base,
+  [Theme.Dark]: palette.gray.dark2,
+};