+++ packages/menu/src/Menu.story.tsx
@@ -5,10 +5,18 @@ import CloudIcon from '@leafygreen-ui/icon/dist/Cloud';
 import EllipsisIcon from '@leafygreen-ui/icon/dist/Ellipsis';
 import IconButton from '@leafygreen-ui/icon-button';
 import LeafyGreenProvider from '@leafygreen-ui/leafygreen-provider';
-import { storybookArgTypes, StoryMeta } from '@leafygreen-ui/lib';
+import { StoryMeta } from '@leafygreen-ui/lib';
 
 import { Size } from './types';
-import { Menu, MenuItem, MenuProps, MenuSeparator, SubMenu } from '.';
+import {
+  Menu,
+  MenuItem,
+  MenuItemProps,
+  MenuProps,
+  MenuSeparator,
+  SubMenu,
+  SubMenuProps,
+} from '.';
 
 export default StoryMeta({
   title: 'Components/Menu',
@@ -16,7 +24,7 @@ export default StoryMeta({
   parameters: {
     default: 'SubMenuExample',
     controls: {
-      exclude: ['trigger', 'children', 'refEl', 'setOpen'],
+      exclude: ['trigger', 'children', 'refEl', 'setOpen', 'as'],
     },
   },
   args: {
@@ -32,7 +40,6 @@ export default StoryMeta({
     usePortal: {
       control: 'boolean',
     },
-    darkMode: storybookArgTypes.darkMode,
     size: {
       options: Object.values(Size),
       control: 'select',
@@ -47,7 +54,7 @@ export const UncontrolledTemplate = ({
   open,
   darkMode,
   ...args
-}: MenuProps & { size: Size }) => {
+}: MenuProps & MenuItemProps & SubMenuProps) => {
   return (
     <LeafyGreenProvider>
       <Menu
@@ -158,7 +165,7 @@ export const Controlled = ({
   trigger,
   darkMode,
   ...args
-}: MenuProps & { size: Size }) => {
+}: MenuProps & MenuItemProps & SubMenuProps) => {
   const [isOpen, setIsOpen] = useState(open);
   return UncontrolledTemplate({
     size,