+++ packages/table/src/Table/Table.types.ts
@@ -21,9 +21,4 @@ export interface TableProps<T extends LGRowData>
    * The `useLeafyGreenTable` return value
    */
   table?: LeafyGreenTable<T>;
-
-  /**
-   * Classname passed to the `<div>` that wraps the Table component
-   */
-  tableContainerClassName?: string;
 }