+++ packages/text-input/src/TextInput/TextInput.styles.ts
@@ -7,6 +7,7 @@ import {
   fontFamilies,
   fontWeights,
   hoverRing,
+  spacing,
   transitionDuration,
   typeScales,
 } from '@leafygreen-ui/tokens';
@@ -47,8 +48,8 @@ export const getWrapperFontSize = (
 
     case SizeVariant.Large: {
       return css`
-        font-size: 18px;
-        line-height: 32px;
+        font-size: ${typeScales.large.fontSize}px;
+        line-height: ${typeScales.large.lineHeight}px;
       `;
     }
 
@@ -72,10 +73,9 @@ export const inheritTypeScale = css`
 `;
 
 export const textContainerStyle = css`
-  > *:last-child {
-    display: block;
-    padding-bottom: 4px;
-  }
+  display: flex;
+  flex-direction: column;
+  margin-bottom: ${spacing[1]}px;
 `;
 
 export const inputContainerStyle = css`
@@ -279,6 +279,38 @@ export const inputSizeStyles: Record<SizeVariant, string> = {
   `,
 };
 
+// Below numbers are "magic" because they need to account for border, font's letter widths, etc
+export const inputPaddingForIndicator: Record<SizeVariant, string> = {
+  [SizeVariant.XSmall]: css`
+    padding-right: 31px;
+  `,
+  [SizeVariant.Small]: css`
+    padding-right: 34px;
+  `,
+  [SizeVariant.Default]: css`
+    padding-right: 37px;
+  `,
+  [SizeVariant.Large]: css`
+    padding-right: 39px;
+  `,
+};
+
+// Below numbers are "magic" because they need to account for border, font's letter widths, etc
+export const inputPaddingForOptionalText: Record<SizeVariant, string> = {
+  [SizeVariant.XSmall]: css`
+    padding-right: 64px;
+  `,
+  [SizeVariant.Small]: css`
+    padding-right: 69px;
+  `,
+  [SizeVariant.Default]: css`
+    padding-right: 71px;
+  `,
+  [SizeVariant.Large]: css`
+    padding-right: 74px;
+  `,
+};
+
 export const inputStateStyles: Record<State, Record<Theme, string>> = {
   [State.Valid]: {
     [Theme.Light]: css`
@@ -347,20 +379,23 @@ export const inputIndicatorStyle = css`
   display: flex;
   align-items: center;
   z-index: 1;
+  padding-left: ${spacing[2]}px;
+  right: 2px; // account for border width
 `;
 
+// Below numbers are "magic" because they need to account for border width, etc that would need larger refactoring to make DRY
 export const inputIndicatorSizeStyle: Record<SizeVariant, string> = {
   [SizeVariant.XSmall]: css`
-    right: 10px;
+    padding-right: 6px;
   `,
   [SizeVariant.Small]: css`
-    right: 10px;
+    padding-right: 10px;
   `,
   [SizeVariant.Default]: css`
-    right: 12px;
+    padding-right: 12px;
   `,
   [SizeVariant.Large]: css`
-    right: 16px;
+    padding-right: 14px;
   `,
 };
 
@@ -388,8 +423,14 @@ export const stateIndicatorStyles: Record<
 
 export const optionalTextBaseStyle = css`
   font-size: 12px;
+  line-height: 12px;
   font-style: italic;
   font-weight: ${fontWeights.regular};
+  display: flex;
+  align-items: center;
+  > p {
+    margin: 0;
+  }
 `;
 
 export const optionalTextThemeStyle: Record<Theme, string> = {