+++ packages/callout/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @leafygreen-ui/callout
 
+## 9.0.7
+
+### Patch Changes
+
+- 3bd3e0c45: Update Storybook for Live Example on .design
+- Updated dependencies [55d33e435]
+- Updated dependencies [07db42330]
+- Updated dependencies [55d33e435]
+- Updated dependencies [cf00160ec]
+- Updated dependencies [111b680c5]
+- Updated dependencies [77320a6b8]
+  - @leafygreen-ui/lib@10.3.3
+  - @leafygreen-ui/palette@4.0.4
+  - @leafygreen-ui/typography@16.3.0
+  - @leafygreen-ui/tokens@2.0.3
+
 ## 9.0.6
 
 ### Patch Changes