+++ packages/menu/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @leafygreen-ui/menu
 
+## 20.0.3
+
+### Patch Changes
+
+- cf00160ec: Updates TSDocs
+- ce0fcb3f6: Excludes `children` from story controls
+- Updated dependencies [55d33e435]
+- Updated dependencies [55d33e435]
+- Updated dependencies [cf00160ec]
+- Updated dependencies [ce0fcb3f6]
+- Updated dependencies [111b680c5]
+- Updated dependencies [77320a6b8]
+  - @leafygreen-ui/lib@10.3.3
+  - @leafygreen-ui/palette@4.0.4
+  - @leafygreen-ui/icon-button@15.0.10
+  - @leafygreen-ui/popover@11.0.8
+  - @leafygreen-ui/tokens@2.0.3
+
 ## 20.0.2
 
 ### Patch Changes