+++ packages/search-input/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @leafygreen-ui/search-input
 
+## 2.0.3
+
+### Patch Changes
+
+- 55d33e435: Update to BaseFontSize prop control for .design live example
+- Updated dependencies [55d33e435]
+- Updated dependencies [77320a6b8]
+- Updated dependencies [07db42330]
+- Updated dependencies [55d33e435]
+- Updated dependencies [cf00160ec]
+- Updated dependencies [ce0fcb3f6]
+- Updated dependencies [111b680c5]
+- Updated dependencies [77320a6b8]
+  - @leafygreen-ui/input-option@1.0.3
+  - @leafygreen-ui/lib@10.3.3
+  - @leafygreen-ui/palette@4.0.4
+  - @leafygreen-ui/typography@16.3.0
+  - @leafygreen-ui/icon-button@15.0.10
+  - @leafygreen-ui/popover@11.0.8
+  - @leafygreen-ui/tokens@2.0.3
+
 ## 2.0.2
 
 ### Patch Changes