+++ packages/table/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @leafygreen-ui/table
 
+## 12.0.0
+
+### Major Changes
+
+- 632f0a461: - API changes to ensure react-table exports are not required in usage
+  - See [upgrade guide](https://github.com/mongodb/leafygreen-ui/blob/main/packages/table/UPGRADE.md) for details
+
+### Patch Changes
+
+- Updated dependencies [55d33e435]
+- Updated dependencies [07db42330]
+- Updated dependencies [55d33e435]
+- Updated dependencies [cf00160ec]
+- Updated dependencies [ce0fcb3f6]
+- Updated dependencies [111b680c5]
+- Updated dependencies [77320a6b8]
+  - @leafygreen-ui/lib@10.3.3
+  - @leafygreen-ui/palette@4.0.4
+  - @leafygreen-ui/typography@16.3.0
+  - @leafygreen-ui/icon-button@15.0.10
+  - @leafygreen-ui/tokens@2.0.3
+
 ## 11.0.3
 
 ### Patch Changes