+++ packages/typography/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @leafygreen-ui/typography
 
+## 16.3.0
+
+### Minor Changes
+
+- 07db42330: Exports typography component types
+
+### Patch Changes
+
+- 55d33e435: Update to BaseFontSize prop control for .design live example
+- cf00160ec: Updates TSDocs
+- Updated dependencies [55d33e435]
+- Updated dependencies [55d33e435]
+- Updated dependencies [cf00160ec]
+- Updated dependencies [111b680c5]
+- Updated dependencies [77320a6b8]
+  - @leafygreen-ui/lib@10.3.3
+  - @leafygreen-ui/palette@4.0.4
+  - @leafygreen-ui/tokens@2.0.3
+
 ## 16.2.1
 
 ### Patch Changes